+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-## 2.0.6 (WIP)
+## 2.1.0 (WIP)
+- Fix to evict the destination file from local cache post rename file operation.
 
 ## 2.0.5 (2023-08-02)
 **Features**
 - In case of MSI based authentication, user shall provide object-id of the identity and honour-acl flag for file-system to work with ACLs assigned to the given identity instead of permissions.
 - Added support to read OAuth token from a user given file.
+- Added support for ARM64 architecture.
 
 **Bug Fixes**
 - Fixed priority level check of components to reject invalid pipeline entries.